--- a/modules/postgres/postgres.go
+++ b/modules/postgres/postgres.go
@@ -19,6 +19,10 @@ const (
 	defaultSnapshotName  = "migrated_template"
 )
 
+// SQLDriverName is passed to sql.Open() to connect to the database when making or restoring snapshots.
+// This can be set if your app imports a different postgres driver, f.ex. "pgx"
+var SQLDriverName = "postgres"
+
 // Container represents the postgres container type used in the module
 type Container struct {
 	*testcontainers.DockerContainer
@@ -269,7 +273,7 @@ func (c *Container) snapshotConnection(ctx context.Context) (*sql.Conn, func(), 
 
 	// Try to use an actual postgres connection, if the driver is loaded
 	connStr := c2.MustConnectionString(ctx, "sslmode=disable")
-	pool, err := sql.Open("postgres", connStr)
+	pool, err := sql.Open(SQLDriverName, connStr)
 	if err != nil {
 		return nil, nil, fmt.Errorf("sql.Open for snapshot connection failed: %w", err)
 	}