<!-- Type of change
Please label this PR with one of the existing labels, depending on the scope of your change
-->

## What does this PR do?

The postgres module has functions for saving snapshots and restoring them later.
This feature is helpful for integration testing without having to manually clear the database.

This PR speeds up the snapshot/restore functionality by executing the necessary SQL commands via a regular `sql.DB` connection.

If the driver is not available in the test environment, or the commands fail for some other reason (SSL config, etc), it falls back to the old variant of calling `docker exec`.

## Why is it important?

`docker exec` can be unnecessarily slow, especially when using docker in a VM with Docker Desktop or colima.

On my laptop (MacBook Air M1), each call to `Container.Restore()` takes about 150~200ms.
In my suite of 20 tests, where each test restores the database before starting, with creating the initial container, this brings the total time up to 12 seconds, which is quite a lot for a few small database tests.

## Related issues

- Relates to https://github.com/testcontainers/testcontainers-go/pull/2199

## How to test this PR

```go
cd modules/postgres
go test ./...
```

## Follow-ups

I've also considered caching the DB pool that is created to manage the databases.
After running some benchmark with this code (from branch X), I noticed it does not have a big impact on the whole test suite - my internal test suite does not change in execution time (since the time is dominated by re-connecting the actual test client, which will be terminated when the database gets deleted during the restore).
The module/postgres test suite even gets slower:

https://github.com/cfstras/testcontainers-go/commit/9ad456dc22ff6f2c8980f61887fb603e82f9f5bd:
```
# At 9ad456dc22ff6f2c8980f61887fb603e82f9f5bd (caching *sql.DB)
➜ cd module/postgres
➜ hyperfine --warmup 1 "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     40.816 s ±  1.697 s    [User: 1.231 s, System: 0.974 s]
  Range (min … max):   38.398 s … 42.658 s    10 runs
```


```
# At e6babe45d6b91c433ce88ae57c6b62bfe46be794 (no caching of *sql.DB)
➜ cd module/postgres
➜ hyperfine --warmup 1 "go test ./... -count 1"
Benchmark 1: go test ./... -count 1
  Time (mean ± σ):     36.602 s ±  1.241 s    [User: 1.166 s, System: 0.953 s]
  Range (min … max):   35.877 s … 39.071 s    10 runs
```