@cfstras 2024-06-20
@mdelapenya
Note that I initially wrote this for `main`, but saw you're working on `v1`, so I rebased. I cannot help but notice that container setup seems a lot slower on this branch - my benchmark wins are almost completely eaten up by this slowdown... Haven't looked further into it yet.