@mdelapenya 2024-06-20
Thanks for submitting this PR, I'll take a look asap.

I'd prefer targeting everything to `main`, as I'm doing the merges/cherries/rebases to avoid the push the burden on the community. What I'm really interested is in how you run those benchmarks. Could you share a bit about it?