@cfstras 2024-06-21
Hi @mdelapenya, thanks for your response :)

I've re-based onto `main`, and ran some of my benchmarks again.
https://github.com/testcontainers/testcontainers-go/pull/2600
Results are posted on the PR.
 
The benchmarks are run with [hyperfine](https://github.com/sharkdp/hyperfine), which is a great tool indeed :)